Fix reference integration requirements.txt
Use a copy of the requirements.txt from the screening reference
integration.

The change is needed because the reference integration was using
Python package Flask version 1.1.1 which requires Jinja2>=2.10.1 but
fails with modern Jinja2 versions 3+, and we were not specifying the
version of Jinja2. There were similar problems with other indirect
dependencies as well.

Really we should switch to using a Pipfile and Pipfile.lock (or
equivalent).

Jira: APPS-391